chore: updated object locations
